Rename all admin handlers to one pattern
