--- handlers/admin/backups/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ("router",)
-
-from .backups_handler import router
