--- handlers/admin/restart/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ("router",)
-
-from .restart_handler import router
